Inline createSmartInput in createField since it's only called there and creates some unnecessary complexity with the module imports
